% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/break_even.R
\docType{data}
\name{break_even}
\alias{break_even}
\title{Detailed run times for analysis of break-even point}
\format{\preformatted{'data.frame':	49200 obs. of  4 variables:
 $ prob: chr  "uniform" "uniform" "uniform" "uniform" ...
 $ expr: Factor w/ 4 levels "crank","rank",..: 3 1 2 3 1 1 1 2 4 2 ...
 $ time: num  2046694 48843 230613 686440 33025 ...
 $ n   : int  100 100 100 100 100 100 100 100 100 100 ...
}}
\usage{
break_even
}
\description{
Run times measured on an Intel(R) Xeon(R) CPU E5-2643 clocked at 3.30 GHz
with 10 MB cache.
}
\keyword{datasets}

