% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_int_R.R
\name{sample_int_R}
\alias{sample_int_R}
\title{Weighted sampling without replacement using built-in routine}
\usage{
sample_int_R(n, size, prob)
}
\arguments{
\item{n}{a positive number, the number of items to choose from.  See
    \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{prob}{A vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
An integer vector of length \code{size} with elements from
  \code{1:n}.
}
\description{
This is a simple wrapper for \code{\link[base]{sample.int}}.
}
\examples{
s <- sample_int_R(2000, 1000, runif(2000))
stopifnot(unique(s) == s)
p <- c(995, rep(1, 5))
n <- 1000
set.seed(42)
stopifnot(abs(table(replicate(sample_int_R(6, 3, p), n=n)) / n -
  c(1, rep(0.4, 5))) < 0.04)
}
\seealso{
\code{\link[base]{sample.int}}
}

