% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timings.R
\docType{data}
\name{timings}
\alias{timings}
\title{Run times}
\format{\preformatted{'data.frame':	9600 obs. of  4 variables:
 $ prob: chr  "uniform" "uniform" "uniform" "uniform" ...
 $ expr: Factor w/ 4 levels "crank","rank",..: 4 2 2 3 1 4 4 2 3 1 ...
 $ time: num  46820 355003 136884 1853908 34963 ...
 $ n   : int  31 31 31 31 31 31 31 31 31 31 ...
}}
\usage{
timings
}
\description{
Run times measured on an Intel(R) Xeon(R) CPU E5-2643 clocked at 3.30 GHz
with 10 MB cache.
}
\keyword{datasets}

