\docType{package}
\name{wrswoR-package}
\alias{wrswoR}
\alias{wrswoR-package}
\title{What the package does (short line)}
\description{
  More about what it does (maybe more than one line) ~~ A
  concise (1-5 lines) description of the package ~~
}
\details{
  \tabular{ll}{ Package: \tab wrswoR\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2013-03-05\cr
  License: \tab What license is it under?\cr } ~~ An
  overview of how to use the package, including the most
  important functions ~~
}
\examples{
\%\% ~~ simple examples of the most important functions ~~
}
\author{
  Who wrote it

  Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
  ~~ Literature or other references for background
  information ~~
}
\seealso{
  ~~ Optional links to other man pages, e.g. ~~ ~~
  \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\keyword{package}

